Checkout SymbolsInput to reuse for both separators & padding symbols
